fix(dashboards): reload drawers on foundation switch and gate headlines (LFXV2-2836)
Address PR #1207 review feedback:
- maintainers/active-contributors/total-projects drawers now react to
 selectedFoundation via combineLatest so distribution/table/lifecycle
 reload when an ED switches foundations with the drawer open
- gate active-contributors MoM headline on momDataLoading(); pass
 reconciledMaintainersData so the maintainers headline hides during a switch
- pin 'en-US' locale on health-scores summary and card tooltip (SSR safety)
- opt events and project-health-scores drawer datasets into zeroStub plugin
- move ZERO_BAR_STUB_PLUGIN to @lfx-one/shared/constants and register once
 via component static state (no module-level constants in apps/lfx-one)
